chore: split session expiry modal into own file
